Account terms only in leaf nodes of type hierarchy
The idea is that all statistics of the abstract nodes are easily
reproducible based on those of their children, in a final aggregation
step. It might even be beneficial to have "unmerged" statistics in
each node that pertain strictly to that node only.

The performance advantage of only modifying one node per term is
obvious.